Wrap MappedTrackInfo
