+++ core/src/main/java/com/novoda/noplayer/exoplayer/mediasource/ExoPlayerTrackSelector.java
@@ -46,27 +46,6 @@ public class ExoPlayerTrackSelector {
         return new ExoPlayerMappedTrackInfo(trackInfo);
     }
 
-    public static class ExoPlayerMappedTrackInfo {
-
-        private final MappingTrackSelector.MappedTrackInfo mappedTrackInfo;
-
-        public ExoPlayerMappedTrackInfo(MappingTrackSelector.MappedTrackInfo mappedTrackInfo) {
-            this.mappedTrackInfo = mappedTrackInfo;
-        }
-
-        public int length() {
-            return mappedTrackInfo.length;
-        }
-
-        public TrackGroupArray getTrackGroups(int index) {
-            return mappedTrackInfo.getTrackGroups(index);
-        }
-
-        public int getAdaptiveSupport(Integer rendererIndex, int groupIndex, boolean includeCapabilitiesExceededTracks) {
-            return mappedTrackInfo.getAdaptiveSupport(rendererIndex, groupIndex, includeCapabilitiesExceededTracks);
-        }
-    }
-
     void setSelectionOverride(TrackType trackType, TrackGroupArray trackGroups, MappingTrackSelector.SelectionOverride selectionOverride) {
         Integer rendererIndex = rendererTrackIndex.get(trackType);
         trackSelector.setSelectionOverride(rendererIndex, trackGroups, selectionOverride);
